Rename some folders to follow java best practices
