--- src/main/java/Utils/IClientValidationHelper.java
@@ -1,9 +1,0 @@
-package utils;
-
-import model.Client;
-
-public interface IClientValidationHelper {
-
-    //validate client fields
-    boolean ValidateClient(Client client);
-}
