+++ src/main/java/model/payments/IPayStrategy.java
