+++ src/main/java/utils/IClientValidationHelper.java
@@ -1,0 +1,9 @@
+package utils;
+
+import model.Client;
+
+public interface IClientValidationHelper {
+
+    //validate client fields
+    boolean ValidateClient(Client client);
+}
