+++ src/main/java/utils/QueriesCreateHelper.java
